ООП классы и интерфейсы. Пункт 7.//ДЗ
