+++ HomeWorks/Bot/IToDoService.cs
@@ -1,0 +1,11 @@
+namespace Bot;
+
+public interface IToDoService
+{
+    IReadOnlyList<ToDoItem> GetAllByUserId(Guid userId);
+    //Возвращает ToDoItem для UserId со статусом Active
+    IReadOnlyList<ToDoItem> GetActiveByUserId(Guid userId);
+    ToDoItem Add(ToDoUser user, string name);
+    void MarkCompleted(Guid id);
+    void Delete(Guid id);
+}
