+++ HomeWorks/Bot/ToDoItem.cs
@@ -4,7 +4,7 @@ public class ToDoItem
 {
     private ToDoItemState _state;
     public Guid Id { get; init; }
-    public User User { get; init; }
+    public ToDoUser ToDoUser { get; init; }
     public string Name { get; init; }
     public DateTime CreatedAt { get; init; }
 
@@ -22,10 +22,10 @@ public class ToDoItem
     }
     public DateTime? StateChangedAt { get; set; }
 
-    public ToDoItem(string name, User user)
+    public ToDoItem(string name, ToDoUser toDoUser)
     {
         Id = Guid.NewGuid();
-        User = user;
+        ToDoUser = toDoUser;
         Name = name;
         CreatedAt = DateTime.Now;
         State = ToDoItemState.Active;
